requirements.txt
matplotlib==3.10.3
matplotlib-inline==0.1.7
numpy==2.3.1
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
xgboost
mlflow
python-dotenv
psycopg2-binary
google-cloud-storage